chore: organize project structure
