--- src/App.css
@@ -1,8 +1,0 @@
-main {
-  left: 154px;
-  overflow-y: auto;
-}
-
-.sidebar {
-  width: 100px;
-}
