+++ src/components/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { MuiThemeProvider } from '@material-ui/core/styles'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import NavTabs from './NavTabs'
-import theme from './theme'
+import theme from '../theme'
 
 export default class NewApp extends Component {
   static displayName = 'App'
