+++ src/components/Nodes/GethConfig/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import styled, { css } from 'styled-components'
 import Typography from '@material-ui/core/Typography'
-import Button from '../../../components/Button'
+import Button from '../../shared/Button'
 import { Mist } from '../../../API'
 import VersionList from './VersionList'
 import ConfigForm from './ConfigForm'
