--- src/containers/normalize.css
@@ -1,28 +1,0 @@
-#root main {
-  background-color: rgb(247, 247, 247);
-  padding: 1rem;
-}
-
-#root main h1 {
-  margin-top: 5px;
-  margin-bottom: 18px;
-}
-
-#root main h2 {
-  margin-bottom: 18px;
-}
-
-
-.card-container {
-  display: grid;
-  padding: 1rem;
-  grid-template-columns: repeat(3, 200px);
-  grid-gap: 1rem;
-}
-
-.card {
-  border-radius: .25rem;
-  box-shadow: 0 0.125rem 0.5rem rgba(34,34,34,.125);
-  margin-bottom: .25rem;
-}
-
