+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
 import { Provider } from 'react-redux'
-import App from './App'
-import Popup from './Popup'
+import App from './components/App'
+import Popup from './components/popups'
 import { Mist } from './API'
 import store from './API/ReduxStore'
